+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `Scratchpad::placement_static_new()` for initializing a `Scratchpad` within
+  a block of uninitialized memory for `Scratchpad` types that use only static
+  arrays for backing memory, bypassing potential call stack use for parameters
+  and return values that may exceed the amount of space available on the
+  stack.
+
 ## [1.0.1] - 2018-06-07
 ### Fixed
 - Use unaligned reads and writes in `Tracking::get()` and `Tracking::set()`
@@ -167,6 +175,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
+[Unreleased]: https://github.com/okready/scratchpad/compare/v1.0.1...HEAD
 [1.0.1]: https://github.com/okready/scratchpad/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/okready/scratchpad/compare/v1.0.0-beta.2...v1.0.0
 [1.0.0-beta.2]: https://github.com/okready/scratchpad/compare/v1.0.0-beta.1...v1.0.0-beta.2